feat(tui): integrate body decoder registry for protobuf/gRPC display
Wire BodyDecoderRegistry through TUI ports so detail view can decode
protobuf and gRPC-Web bodies into human-readable text. Includes:

- Registry wiring in main.go with proto file loading
- renderOpts refactor to thread decoder through detail pipeline
- Remove protobuf MIME types from binary blocklist (now decoded)
- E2E harness support for body decoder + gRPC Connect test
- Fix gosec G115: safe uint64 cast in gRPC-Web frame parser
- Makefile generate-proto target for buf codegen

+++ internal/bodydecoder/grpcweb.go
@@ -56,10 +56,10 @@ func (d *GRPCWebDecoder) Decode(body []byte, meta DecoderMetadata) (string, stri
 }
 
 const (
-	frameHeaderLen    = 5
-	flagData          = 0x00
-	flagCompressed    = 0x01
-	flagTrailers      = 0x80
+	frameHeaderLen = 5
+	flagData       = 0x00
+	flagCompressed = 0x01
+	flagTrailers   = 0x80
 )
 
 // extractDataFrames parses gRPC-Web framing and returns the concatenated
@@ -77,7 +77,7 @@ func extractDataFrames(body []byte) (payload []byte, notes []string, err error) 
 
 		truncated := false
 		frameData := remaining
-		if uint32(len(remaining)) < length {
+		if uint64(len(remaining)) < uint64(length) {
 			// Truncated frame — decode what we have.
 			truncated = true
 			notes = append(notes, fmt.Sprintf("truncated frame: expected %d bytes, got %d", length, len(remaining)))
